Provide DB-specific locations through conditional compilation
Use a single data structure, `databaseLocation`, to describe either a work directory or in-memory when compiled for SQLite
or a remote URL when compiled for PostgreSQL. This provides strong type safety but comes at the cost of having to include a
argument parser in the DB modules, which is not really where it belongs. Alternatively, conditional compilation could be
added to the argument parsing module, but that also seems out of place. Ultimately, including all 3 alternatives in
`databaseLocation` and validating at runtime may be a cleaner solution, despite the loss of type safety.