Create remote URL option for DB location
Add a command line option to specify a remote URL as DB location instead of work directory
or in memory. This was added as another sum type to the existing option to prevent additional
conditional compilation and avoid the need to conditionally compile the option parser. This
comes at the expense of type safety, as each DB module throws a runtime error if an
incompatible DB Location is provided.

Trace logging of the max concurrency configuration was kept in `Kupo` rather than the DB modules
due to a cyclical dependency between `App/Configuration`, `SQLite`/`Postgres`, and `App/Database`.