Replace `newDBPoolFromFile` with `newDBPool`
Including `newDBPoolFromFile` in the external API of `App/Database`
creates issues for the Postgres module, which does not have any
notion of a DB file. Instead, we add a `FilePath` to
`Configuration.InMemory` but do not expose ability to provide a
filepath to the command line. This allows us to provide a filepath
for testing purposes but does allow the user to supply one, all
while maintaining good abstraction boundaries.