Add an '--until' option
Sometimes it is useful to sync only up to a specific point. For example,
suppose you want to get a snapshot of the wallet balances for some token
at a single point in time, such as to tally an on-chain vote or poll.

You could index without --prune-utxo, but this may be a very expensive
syncing operation.

However, if you index *with* --prune-utxo, then as soon as kupo passes
the snapshot point, you run the risk of pruning UTxOs that were spent
after the snapshot, which would distort the snapshot you obtain.

This commit adds an --until option.

Kupo will index as normal, until it reaches the specified point,
after which it will stop consuming blocks and continue to answer
http queries about the data.

To do this, I pulled the `rollForward` function out, so the consumer
could be given some behavior, and then provided two options:
- rollForward, which ehaves as normal
- rollForwardUntil, which executes rollForward on all blocks before the
  point

Note: `--until` is currently exclusive: it will roll *up to* but
*not include* the point you specify. It is unclear what the most
intuitive semantics should be, so I'm happy to change this. It doesn't
particularly matter for my use case.