Fix Hydra integration w/ 0.20.0 & 0.21.0
Update decoding of SnapshotConfirmed to use the included transactions
directly. This is not needed anymore as the API changed.

Also, the L2 ledger of hydra is always in ConwayEra now.

There is now also a golden file for each server output that the
hydra-node sends through the websocket that we can use in tests.